Fix: add more type checks on quiz controller

+++ api/controllers/QuizController.js
@@ -27,12 +27,19 @@ module.exports = {
 
 		let data = _.pick(req.body, allowedParamters);
 		data.author = req.current_user;
-		QuizService.create(data, (err, createdQuiz) => {
-			if(err) return res.badRequest(err);
-			if (createdQuiz.hasOwnProperty('title')) {
-				return res.created(createdQuiz);
+		QuestionService.listOne(data.question, (err, foundQuestion) => {
+			if(err) return res.notFound(err);
+			if(typeof foundQuestion != 'undefined' && foundQuestion.hasOwnProperty('title')) {
+				QuizService.create(data, (err, createdQuiz) => {
+					if(err) return res.badRequest(err);
+					if (typeof createdQuiz != 'undefined' && createdQuiz.hasOwnProperty('title')) {
+						return res.created(createdQuiz);
+					} else {
+						return res.serverError('Internal Server Error');
+					}
+				})
 			} else {
-				return res.serverError('Internal Server Error');
+				return res.notFound('Question does not exist');
 			}
 		})
 	},
@@ -48,7 +55,7 @@ module.exports = {
 		data.id = req.param.id;
 		QuizService.listOne(id, (err, foundQuiz) => {
 			if(err) return res.notFound(err);
-			if(foundQuiz.hasOwnProperty('title')) {
+			if(typeof foundQuiz != 'undefined' && foundQuiz.hasOwnProperty('title')) {
 				QuizService.update(data, (err, updatedQuiz) => {
 					if(err) return res.badRequest(err);
 					if (updatedQuiz.hasOwnProperty('title')) {
@@ -64,8 +71,8 @@ module.exports = {
 	destroy: (req, res) => {
 		let user = req.param('id');
         QuizService.listOne(user, (err, foundQuiz) => {
-            if (err) return res.notFound('Question not found');
-            if (foundQuiz.hasOwnProperty('title')) {
+            if (err) return res.notFound('Quiz not found');
+            if (typeof foundQuiz != 'undefined' && foundQuiz.hasOwnProperty('title')) {
                 QuizService.destroy(user, (err, quiz) => {
                     if(err) return res.serverError('Internal Server Error');
                     if(quiz) {
